Problem: SumRootToLeafNumbers Problem Description: Given a binary tree containing digits from 0-9 only, each root-to-leaf path could represent a number.
An example is the root-to-leaf path 1->2->3 which represents the number 123.

Find the total sum of all root-to-leaf numbers.

For example,

    1
   / \
  2   3
The root-to-leaf path 1->2 represents the number 12.
The root-to-leaf path 1->3 represents the number 13.

Return the sum = 12 + 13 = 25.

Has bugs after 30 min test, but able to fix it soon.
Two bugs: 1. Line 46: pop_back(): (used to put it into the else{} )
2. line 66 result = (result + sum[i])>=2147483647?2147483647:(result);